backend/requirements.txt
flask
flask-cors
requests
pandas
scikit-learn
python-dotenv
numpy
deep-translator
